worker/workerMetrics/registry.go
package workerMetrics

import (
	"fmt"
	"regexp"
	"sync"
)

// Registry is what can create metrics
type Registry struct ***REMOVED***
	metrics map[string]*Metric
	l       sync.RWMutex
***REMOVED***

// NewRegistry returns a new registry
func NewRegistry() *Registry ***REMOVED***
	return &Registry***REMOVED***
		metrics: make(map[string]*Metric),
	***REMOVED***
***REMOVED***

const nameRegexString = "^[\\p***REMOVED***L***REMOVED***\\p***REMOVED***N***REMOVED***\\._ !\\?/&#\\(\\)<>%-]***REMOVED***1,128***REMOVED***$"

var compileNameRegex = regexp.MustCompile(nameRegexString)

func checkName(name string) bool ***REMOVED***
	return compileNameRegex.Match([]byte(name))
***REMOVED***

// NewMetric returns new metric registered to this registry
// TODO have multiple versions returning specific metric types when we have such things
func (r *Registry) NewMetric(name string, typ MetricType, t ...ValueType) (*Metric, error) ***REMOVED***
	r.l.Lock()
	defer r.l.Unlock()

	if !checkName(name) ***REMOVED***
		return nil, fmt.Errorf("Invalid metric name: '%s'", name) //nolint:golint,stylecheck
	***REMOVED***
	oldMetric, ok := r.metrics[name]

	if !ok ***REMOVED***
		m := newMetric(name, typ, t...)
		r.metrics[name] = m
		return m, nil
	***REMOVED***
	if oldMetric.Type != typ ***REMOVED***
		return nil, fmt.Errorf("metric '%s' already exists but with type %s, instead of %s", name, oldMetric.Type, typ)
	***REMOVED***
	if len(t) > 0 ***REMOVED***
		if t[0] != oldMetric.Contains ***REMOVED***
			return nil, fmt.Errorf("metric '%s' already exists but with a value type %s, instead of %s",
				name, oldMetric.Contains, t[0])
		***REMOVED***
	***REMOVED***
	return oldMetric, nil
***REMOVED***

// MustNewMetric is like NewMetric, but will panic if there is an error
func (r *Registry) MustNewMetric(name string, typ MetricType, t ...ValueType) *Metric ***REMOVED***
	m, err := r.NewMetric(name, typ, t...)
	if err != nil ***REMOVED***
		panic(err)
	***REMOVED***
	return m
***REMOVED***

// Get returns the Metric with the given name. If that metric doesn't exist,
// Get() will return a nil value.
func (r *Registry) Get(name string) *Metric ***REMOVED***
	return r.metrics[name]
***REMOVED***
